+++ 3501Project/src/AssetManager.h
@@ -43,7 +43,7 @@ namespace AsteroidGame{
 
 		void init(Ogre::SceneManager*);
 		
-		Ogre::SceneNode* createAsteroid(double aRadius);
+		Ogre::SceneNode* createAsteroid(float aRadius);
 		Ogre::SceneNode* createProjectile(long counter, int type);
 		Ogre::SceneNode* createBoundingBox(Ogre::Vector3 aDimensions);
 		Ogre::SceneNode* getPlayerChassis() { return iPlayerChassis; }
